pyproject.toml
[project]
name = "telegram-color-clicker-game-bot"
version = "0.0.1"
description = "Color Clicker: Pick the correctly labeled color fast—before time runs out. How long can you keep up?"
authors = [
    { name = "Ben Carver", email = "benrcarver@gmail.com" }
]
readme = "README.md"
license = "MIT"
license-files = [
    "LICENSE",
]
requires-python = ">=3.8"

dependencies = [
    "python-telegram-bot>=20.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "yt-dlp>=2024.1.1",
    "python-dotenv==0.9.9",
    "flask"
]

[tool.setuptools]
packages = ["telegram_color_clicker_game_bot"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
